requirements.txt
alembic==1.13.3
blinker==1.8.2
Brotli==1.1.0
cachelib==0.13.0
cffi==1.17.1
chardet==5.2.0
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
defusedxml==0.7.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.0
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.2
msgspec==0.18.6
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
six==1.16.0
SQLAlchemy==2.0.36
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
webencodings==0.5.1
Werkzeug==3.0.4
zopfli==0.2.3.post1
